Migrate environment to engine module
